feature: add rpc client
